+++ test.py
@@ -1,7 +1,11 @@
 
+from collections import defaultdict
+import numpy as np
+
 from src import isovec as iso
 
 # TODO: copy function
 # TODO (far): IAEA Livechart Data Download API class
+
 
 pass